requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.3
astroid==2.3.3
beautifulsoup4==4.8.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.2
django-allauth==0.41.0
django-appconf==1.0.4
django-bootstrap4==1.1.1
django-cors-headers==3.7.0
django-crispy-forms==1.8.1
django-imagekit==4.0.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
future==0.18.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pefile==2019.4.18
pilkit==2.0
Pillow==7.1.1
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
pygame==1.9.6
PyInstaller==3.6
PyJWT==2.0.1
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.9
virtualenv==20.4.0
virtualenv-clone==0.5.4
wrapt==1.11.2
